requirements.txt
torch
pandas
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
kaggle
mysql-connector-python
torchtext
torchvision
transformers
dvc
-e .